BipedalWalker_PolicyGradient/requirements.txt
torch>=2.0.0,<3.0.0
gymnasium[box2d]>=0.29.0,<1.0.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
pyyaml>=6.0,<7.0.0
